// src/app/easy/models/Report.ts
export class Report {
    confirm_from_date: Date;
    confirm_to_date: Date;
    arrival_from_date: Date;
    arrival_to_date: Date;
    departure_from_date: Date;
    departure_to_date: Date;
    branch_code: string;
    priority_code: string;
    vendor_code: string;
    status_code: string;
    service_code: string;
    birth_month: string;
}